Reworking Movement Tree
Added and then removed panic state. The flee state will main is as all panic was doing was adding more erratic behavior. This seems like it can still be achieved with a simple isPanicking bool. Less moving parts between states will hopefully make expanding the tree easier as I fear adding more states with two very similar states in the shuffle currently that the resulting bugs will be time consuming t patch and require a large amount of trial and error.

Small side note is the crude vision trigger is functioning properly which is a first with anything iv've written. No more will mobs stumble around and have to bump/sniff another mob to detect their faction and if they should turn tail. Eventually i want move away from using tags to differentiate as I plan to have the number of factions grow over time and for their relationships to be dynamic.

But that will all come in time. For now an operational demo is the goal and I intend to have one within a month Unity permitting.